% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_5primeutr.R
\name{add_five_prime_utr}
\alias{add_five_prime_utr}
\title{Add Five Prime UTR Annotations Based on CAGE Peaks}
\usage{
add_five_prime_utr(annotation_gr, cage_peaks, dist = 500)
}
\arguments{
\item{annotation_gr}{A \code{GRanges} object containing genomic annotations with columns for \code{type}, \code{transcript_id}, and
strand information. It should include annotations for transcripts, genes, and exons.}

\item{cage_peaks}{A \code{GRanges} object containing CAGE peaks, with the \code{thick} column representing the peak regions.}

\item{dist}{The maximum distance upstream of the transcription start site (TSS) within which to search for CAGE peaks.
Default is 500 base pairs.}
}
\value{
A \code{GRanges} object containing the original genomic annotations plus the newly added five prime UTR annotations
and adjusted transcript and exon start positions based on the nearest upstream CAGE peaks within the specified distance.
}
\description{
This function enhances genomic annotations by identifying the closest upstream CAGE peak within a specified distance
to the transcription start site (TSS) of each transcript. It adjusts transcript and exon starts and adds five prime UTR
annotations accordingly. The function operates on a \code{GRanges} object containing genomic annotations and another \code{GRanges}
object representing CAGE peaks.
}
\examples{
# Assuming `annotation_gr` is your GRanges object with transcript, gene, and exon annotations,
# and `cage_peaks` is another GRanges object with CAGE peak data:
annotation_gr <- system.file("extdata", "OKI.I69_v2_minimized.gff3", package="annotationpolish") |> rtracklayer::import()
cage_peaks <- read_cage("https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_no_OKItoOKI.bed", "https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_sl_OKItoOKI.bed" )
updated_annotations <- add_five_prime_utr(annotation_gr, cage_peaks, dist = 500)

# This will return a GRanges object with adjusted annotations and added five prime UTRs.
}
