% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_3primeutr.R
\name{add_three_prime_utr_wrapper}
\alias{add_three_prime_utr_wrapper}
\title{Extend 3' UTRs to Nearest Downstream Motif}
\usage{
add_three_prime_utr_wrapper(
  annotation_gr,
  genome,
  motif = "AATAAA",
  dist = 500
)
}
\arguments{
\item{annotation_gr}{A \code{GRanges} object containing the genomic annotations, typically including transcripts for which 3' UTRs need to be extended. This object should be the result of importing genomic annotation data, such as from a GFF3 file.}

\item{genome}{A \code{BSgenome} object representing the genome to be scanned for the specified motif. This object is expected to encapsulate the entire genomic sequence for the organism of interest, facilitating the search for motif occurrences.}

\item{motif}{A character string representing the DNA sequence motif to search for within the genome as potential 3' UTR extension points. Default is "AATAAA", a common polyadenylation signal motif.}

\item{dist}{An integer specifying the maximum distance (in base pairs) to search downstream of each transcript's 3' end for the motif. The function will only consider motifs within this distance for extending the 3' UTR. Default value is 500 bp.}
}
\value{
The original \code{GRanges} object (annotation_gr) with modified transcript models, where 3' UTRs have been potentially extended to the nearest downstream motif within the specified distance. The function does not explicitly return the object; ensure to capture the result into a variable.
}
\description{
This function extends the 3' untranslated regions (UTRs) of transcripts in a given genomic annotation to the nearest downstream instance of a specified motif, within a maximum specified distance. It first scans the provided genome for occurrences of the motif to create a set of potential UTR extension points. Then, it adjusts the transcript models in the annotation, extending the 3' UTRs to these points where applicable.
}
\examples{
# Assuming 'annotation_gr' is loaded from a GFF3 file and 'BSgenome.Oidioi.OIST.OKI2018.I69' is loaded
updated_annotations <- add_three_prime_utr_wrapper(annotation_gr, BSgenome.Oidioi.OIST.OKI2018.I69, motif = "AATAAA", dist = 500)
# 'updated_annotations' now contains transcripts with potentially extended 3' UTRs.
}
