% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_operons.R
\name{calcOperons.onestrand}
\alias{calcOperons.onestrand}
\title{Calculate Operons on a Single Strand}
\usage{
calcOperons.onestrand(genes, window = 500)
}
\arguments{
\item{genes}{A \code{GRanges} object containing genomic coordinates of genes on a single strand.}

\item{window}{The maximum distance between consecutive genes in base pairs to consider them as part of the same operon. Default is 500 bp.}
}
\value{
A \code{GRanges} object representing the genomic coordinates of identified operons on the given strand.
}
\description{
Helper function for \code{calcOperons} to calculate operons on a single genomic strand. This is used internally by \code{calcOperons} to handle plus and minus strands separately.
}
