% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_operons.R
\name{calcOperons}
\alias{calcOperons}
\title{Calculate Operons}
\usage{
calcOperons(genes, window = 500)
}
\arguments{
\item{genes}{A \code{GRanges} object containing genomic coordinates of genes.}

\item{window}{The maximum distance between consecutive genes in base pairs to consider them as part of the same operon. Default is 500 bp.}
}
\value{
A \code{GRanges} object representing the genomic coordinates of identified operons.
}
\description{
Calculates operons based on genomic coordinates of genes. Operons are identified by grouping genes on the same strand that are within a specified distance from each other.
}
