% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unique_ID.R
\name{check_unique_ids}
\alias{check_unique_ids}
\title{Check for Unique IDs in a GRanges Object}
\usage{
check_unique_ids(gr)
}
\arguments{
\item{gr}{A GRanges object that has been processed to include an 'ID' column with
identifiers for genomic features.}
}
\value{
Returns \code{TRUE} if all IDs in the 'ID' column are unique; otherwise, it returns
\code{FALSE} and emits a warning indicating that some IDs are not unique.
}
\description{
This function checks if the 'ID' values in a GRanges object are unique. It is useful
for ensuring data integrity after assigning unique identifiers to genomic features.
}
\examples{
# Assuming 'gr' is a GRanges object with an 'ID' column
result <- check_unique_ids(gr)
if (result) {
  message("IDs validation passed.")
} else {
  warning("Duplicate IDs found.")
}
}
