% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_manual_annotation.R
\name{clean_manual_anno}
\alias{clean_manual_anno}
\title{Clean Manual Annotations}
\usage{
clean_manual_anno(file)
}
\arguments{
\item{file}{A character string specifying the path to the GFF3 file containing
manual annotation data.}
}
\value{
An object of class \code{GRanges} representing the cleaned manual annotation data.
This object includes selected fields relevant for further analysis, with 'tss_type'
included only if it was present in the original data.
}
\description{
This function imports manual annotation data from a specified GFF3 file. It processes
the annotations to select specific fields, including 'source', 'type', 'score', 'phase',
'Name', 'gene_name', 'gene_id', 'transcript_id', 'Parent', and optionally 'tss_type',
if present. The 'ID' field is explicitly removed to clean the annotation data for further
analysis. Terminal exon and gene/transcript boundaries are adjusted based on the terminal features.
}
\examples{
# Load and clean manual annotation data
cleaned_data <- clean_manual_anno(system.file("extdata", "manual_data.gff3", package="annotationpolish"))

}
