% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_manual_annotation.R
\name{overlaps_within}
\alias{overlaps_within}
\title{Identifies Overlapping Genomic Ranges}
\usage{
overlaps_within(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object representing genomic ranges, where each range specifies a region on a genome.}
}
\value{
A \code{GRanges} object containing only those ranges that overlap with others, excluding any self-overlaps.
}
\description{
This function finds and returns genomic ranges within a \code{GRanges} object that overlap with other ranges,
excluding self-overlaps. It's designed to work with genomic range objects defined in the GenomicRanges package,
leveraging the \code{findOverlaps} method for the identification of overlaps.
}
\examples{
# Assuming `gr` is a GRanges object with genomic ranges:
overlapping_granges <- overlaps_within(gr)
# `overlapping_granges` will contain genomic ranges from `gr` that overlap with others.
}
