% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cage.R
\name{read_cage}
\alias{read_cage}
\title{Read CAGE (Cap Analysis Gene Expression) Data}
\usage{
read_cage(no_sl, sl = NULL)
}
\arguments{
\item{no_sl}{A character string specifying the path or URL to the BED file containing
CAGE data without SL sequences.}

\item{sl}{An optional character string specifying the path or URL to the BED file containing
CAGE data with SL sequences. If provided, data from this file will be combined with the
non-SL data, and each sequence will be annotated with its type.}
}
\value{
An object of class \code{GRanges} representing the imported CAGE data. If both SL and
non-SL data are provided, the return value will be a combined \code{GRanges} object with an
additional column \code{tss_type} indicating whether a sequence is 'sl' (with Spliced Leader)
or 'nosl' (without Spliced Leader).
}
\description{
This function imports CAGE data from provided BED files. It supports importing data
with or without SL (Spliced Leader) sequences. For files with SL sequences, the function
can combine and annotate the two types of sequences (SL and non-SL) for comprehensive analysis.
}
\examples{
no_sl_url <- "https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_no_OKItoOKI.bed"
sl_url <- "https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_sl_OKItoOKI.bed"

# Reading non-SL CAGE data only
read_cage(no_sl_url)

# Reading and combining non-SL and SL CAGE data
read_cage(no_sl_url, sl_url)

}
