% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_utr.R
\name{remove_utr}
\alias{remove_utr}
\title{Remove UTRs and Adjust Transcript and Gene Boundaries}
\usage{
remove_utr(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object containing genomic annotations with at least the columns 'type' and 'transcript_id'.
The 'type' column should indicate the annotation type (e.g., "utr", "transcript", "CDS"). The function
identifies UTRs by detecting 'utr' in the 'type' field, removes these annotations, and then adjusts
the start and end positions of transcripts and genes based on the boundaries of the remaining annotations.}
}
\value{
A \code{GRanges} object similar to the input but with UTR annotations removed and transcript as well as
gene boundaries adjusted based on the remaining annotations. If no UTRs are found, it returns the original \code{GRanges} object.
}
\description{
This function takes a \code{GRanges} object containing genomic annotations and removes UTR annotations from it.
For transcripts associated with UTRs, it adjusts their start and end positions based on the boundaries of the CDS.
The function aims to retain only the coding sequences (CDS) and other relevant annotations while discarding UTRs
and adjusting the genomic ranges of affected transcripts and genes accordingly.
}
\examples{
# Assuming `gr` is your GRanges object with genomic annotations:
gr_updated <- remove_utr(gr)

# This will return a GRanges object with UTRs removed and adjusted annotations.
}
