% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_3primeutr.R
\name{scan_motif}
\alias{scan_motif}
\title{Scan for Motif Occurrences in a Genome}
\usage{
scan_motif(genome, motif = "AATAAA", ...)
}
\arguments{
\item{genome}{A \code{BSgenome} object representing the genome to be scanned. This object should be loaded and passed to the function. The genome object must be from the \code{BSgenome} package, which provides comprehensive genomic sequences for various organisms.}

\item{motif}{A character string representing the DNA motif to be searched for. The default motif is "AATAAA", a common polyadenylation signal in many genomes. The function will also search for the reverse complement of this motif.}
}
\value{
A \code{GRanges} object containing the genomic locations of all occurrences of the specified motif and its reverse complement. Each location is annotated with the chromosome (seqnames), start and end positions (ranges), and the strand ("+" for the forward strand and "-" for the reverse strand) on which the motif occurs.
}
\description{
This function searches for all occurrences of a specified DNA motif and its reverse complement across all chromosomes of a provided genome. It utilizes the \code{BSgenome} package to access genomic sequences and the \code{Biostrings} package for pattern matching. The function is designed to be strand-aware, identifying motif occurrences on both the forward and reverse strands.
}
\examples{
# Load the BSgenome object for a specific organism (Example: Okinawa O. dioica)
genome <- BSgenome::getBSgenome("BSgenome.Oidioi.OIST.OKI2018.I69")

# Scan for the "AATAAA" motif and its reverse complement in the human genome
motif_occurrences <- scan_motif(genome, motif = "AATAAA")
print(motif_occurrences)
}
