% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_operons.R
\name{split_operons_assign_genes}
\alias{split_operons_assign_genes}
\title{Split Operons and Assign Genes Based on CAGE Peaks}
\usage{
split_operons_assign_genes(operons, genes, cage)
}
\arguments{
\item{operons}{A \code{GRanges} object representing operons.}

\item{genes}{A \code{GRanges} object representing genes.}

\item{cage}{A \code{GRanges} object representing CAGE peaks with thick metadata.}
}
\value{
A modified \code{GRanges} object of genes, with operon assignments split using CAGE peaks.
}
\description{
Splits operons and assigns genes based on CAGE peaks, specifically excluding operons overlapping with intergenic CAGE peaks.
}
