% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_transcript.R
\name{split_transcript}
\alias{split_transcript}
\title{Process Genomic Annotations with CAGE Peaks and Gene Size Filtering}
\usage{
split_transcript(
  original_annotation,
  liftoff,
  cage_peaks,
  min_gene_size = 1000,
  remove_no_ORF = F,
  cov_threshold = 0.5
)
}
\arguments{
\item{original_annotation}{A GenomicRanges object containing the original gene annotations.}

\item{liftoff}{A GenomicRanges object with 'liftoff' gene models to be integrated into the original annotation.}

\item{cage_peaks}{A GenomicRanges object containing CAGE peaks used for identifying genes to be split.}

\item{min_gene_size}{Integer, minimum gene size (in base pairs) for a gene to be retained. Default is 1000 bp.}

\item{remove_no_ORF}{Logical, indicating whether to remove liftoff gene models without a valid ORF. Default is FALSE.}

\item{cov_threshold}{Numeric value setting the minimum coverage threshold for 'liftoff' gene models. Used to filter out low-confidence models before integration. Default is 0.5.}
}
\value{
A GenomicRanges object containing the modified gene annotations, with some genes split or replaced based on the overlap with CAGE peaks and the size threshold.
}
\description{
This function refines gene annotations by integrating 'liftoff' gene models with existing annotations, utilizing CAGE peak data to identify and split genes, and filtering genes based on size and optionally the presence of an ORF. It prioritizes genes that intersect with CAGE peaks and meet a size threshold, supporting the identification of potentially novel or modified gene models.
}
\examples{
cage_peaks <- read_cage("https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_no_OKItoOKI.bed", "https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_sl_OKItoOKI.bed" )
original_annotation <- system.file("extdata", "split_example_target.gff3", package="annotationpolish") |> rtracklayer::import()
liftoff <- system.file("extdata", "split_example_liftoff.gff3", package="annotationpolish") |> rtracklayer::import() |> fix_liftoff_oikobase()

split_transcript(original_annotation, liftoff, cage_peaks)
}
