% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_transcript.R
\name{split_transcript}
\alias{split_transcript}
\title{Title}
\usage{
split_transcript(
  original_annotation,
  liftoff,
  cage_peaks,
  min_gene_size = 1000,
  remove_no_ORF = F
)
}
\arguments{
\item{original_annotation}{A GenomicRanges object containing the original gene annotations.}

\item{liftoff}{A GenomicRanges object with 'liftoff' gene models to be integrated into the original annotation.}

\item{cage_peaks}{A GenomicRanges object containing CAGE peaks used for identifying genes to be split.}

\item{min_gene_size}{Integer, minimum gene size (in base pairs) for a gene to be retained. Default is 1000 bp.}

\item{remove_no_ORF}{Logical, indicating whether to remove liftoff gene models without a valid ORF. Default is FALSE.}
}
\value{
A GenomicRanges object containing the modified gene annotations, with some genes split or replaced based on the overlap with CAGE peaks and the size threshold.
}
\description{
This function processes genomic annotations by splitting genes based on their overlap with provided CAGE peaks and filtering based on gene size. It is designed to integrate new gene models from a 'liftoff' annotation with an original annotation, prioritizing genes that intersect with CAGE peaks and are above a minimum size threshold.
}
\examples{
cage_peaks <- read_cage("https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_no_OKItoOKI.bed", "https://raw.githubusercontent.com/oist/LuscombeU-CAGE_libraries/main/2021-12-17_Okinawa_Oik/CrossAlignments/consensus_clusters_sl_OKItoOKI.bed" )
original_annotation <- system.file("extdata", "split_example_target.gff3", package="annotationpolish") |> rtracklayer::import()
liftoff <- system.file("extdata", "split_example_liftoff.gff3", package="annotationpolish") |> rtracklayer::import() |> fix_liftoff_oikobase()

split_transcript(original_annotation, liftoff, cage_peaks)
}
